+++ src/render/OpenGL.c
@@ -324,7 +324,7 @@ static void viewport(int x, int y, int width, int height)
 /* \brief pass interleaved vertex data to OpenGL nicely. */
 static inline void geometryPointer(const glhckGeometry *geometry)
 {
-   // printf("%s (%zu) : %.0f\n", glhckVertexTypeString(geometry->vertexType), geometry->vertexCount, geometry->textureRange);
+   // printf("%s (%zu) : %u\n", glhckVertexTypeString(geometry->vertexType), geometry->vertexCount, geometry->textureRange);
 
    /* vertex data */
    switch (geometry->vertexType) {
@@ -638,7 +638,7 @@ static inline void objectStart(const _glhckObject *object) {
    /* set texture coords according to how geometry wants them */
    GL_CALL(glMatrixMode(GL_TEXTURE));
    GL_CALL(glLoadIdentity());
-   GL_CALL(glScalef(1.0f/object->geometry->textureRange, 1.0f/object->geometry->textureRange, 1.0f));
+   GL_CALL(glScalef((float)1.0f/object->geometry->textureRange, (float)1.0f/object->geometry->textureRange, 1.0f));
 
    /* load view matrix */
    GL_CALL(glMatrixMode(GL_MODELVIEW));
@@ -754,7 +754,7 @@ static inline void textDraw(const _glhckText *text)
 
    GL_CALL(glMatrixMode(GL_TEXTURE));
    GL_CALL(glLoadIdentity());
-   GL_CALL(glScalef(1.0f/text->textureRange, 1.0f/text->textureRange, 1.0f));
+   GL_CALL(glScalef((float)1.0f/text->textureRange, (float)1.0f/text->textureRange, 1.0f));
 
    GL_CALL(glMatrixMode(GL_MODELVIEW));
    GL_CALL(glLoadIdentity());
